moving data out of the reducer
