+++ recipes/src/App.js
@@ -8,13 +8,17 @@ function App() {
   );
 }
 
-
+const mapStateToProps = (state) => {
+  return {
+    
+  }
+};
 
 const mapDispatchToProps = dispatch => {
   return {
-    onListRicipes: () => dispatch({type: 'LISTRICIPES'}),
-    onViewDetails: () => dispatch({type: 'VIEWDETAILS'}),
-    onAddRecipe: () => dispatch({type: 'ADDRICIPE'})
+    onListRicipes: () => dispatch({type: 'LIST_RICIPES'}),
+    onViewDetails: () => dispatch({type: 'VIEW_DETAILS'}),
+    onAddRecipe: () => dispatch({type: 'ADD_RICIPE'})
   }
 };
